Separate ResearchAccount classes from tally code
