+++ src/neuralnet/tally.h
@@ -1,205 +1,12 @@
 #pragma once
 
-#include <memory>
-#include <set>
-
-#include <boost/optional.hpp>
+#include "neuralnet/account.h"
 
 class CBlockIndex;
 
 namespace NN {
 
 class Cpid;
-
-//!
-//! \brief An optional type that contains a pointer to a block index object or
-//! does not.
-//!
-typedef boost::optional<const CBlockIndex*> BlockPtrOption;
-
-//!
-//! \brief Compares CBlockIndex objects by comparing block heights.
-//!
-struct BlockIndexHeightComparator
-{
-    //!
-    //! \brief Determine whether \p a has a height less than \p b.
-    //!
-    //! \return \c true if the height of \p a is less than the height of \p b.
-    //!
-    bool operator()(const CBlockIndex* const a, const CBlockIndex* const b) const
-    {
-        return a->nHeight < b->nHeight;
-    }
-};
-
-//!
-//! \brief Stores the research reward context for a CPID used to calculate
-//! accrued research rewards.
-//!
-class ResearchAccount
-{
-public:
-    int64_t m_total_research_subsidy;    //!< Total lifetime research paid.
-
-    uint16_t m_magnitude;                //!< Current magnitude in superblock.
-    uint32_t m_total_magnitude;          //!< Total lifetime magnitude sum.
-    uint32_t m_accuracy;                 //!< Non-zero magnitude payment count.
-
-    const CBlockIndex* m_first_block_ptr; //!< First block with research reward.
-    const CBlockIndex* m_last_block_ptr; //!< Last block with research reward.
-
-    //!
-    //! \brief Initialize an empty research account.
-    //!
-    ResearchAccount();
-
-    //!
-    //! \brief Determine whether this is a new research account.
-    //!
-    //! \return \c true if no block contains a reward for the account's CPID.
-    //!
-    bool IsNew() const;
-
-    //!
-    //! \brief Get a pointer to the first block with a research reward earned by
-    //! the account.
-    //!
-    //! \return A pointer to the block index if the account earned a research
-    //! reward before.
-    //!
-    BlockPtrOption FirstRewardBlock() const;
-
-    //!
-    //! \brief Get the hash of the first block with a research reward earned by
-    //! the account.
-    //!
-    //! \return The SHA256 hash of the account's first reward block.
-    //!
-    uint256 FirstRewardBlockHash() const;
-
-    //!
-    //! \brief Get the height of the first block with a research reward earned
-    //! by the account.
-    //!
-    //! \return A block height of zero if the account never earned a research
-    //! reward before.
-    //!
-    uint32_t FirstRewardHeight() const;
-
-    //!
-    //! \brief Get the timestamp of the first block with a research reward
-    //! earned by the account.
-    //!
-    //! \return A timestamp of zero if the account never earned a research
-    //! reward before.
-    //!
-    int64_t FirstRewardTime() const;
-
-    //!
-    //! \brief Get a pointer to the last block with a research reward earned by
-    //! the account.
-    //!
-    //! \return A pointer to the block index if the account earned a research
-    //! reward before.
-    //!
-    BlockPtrOption LastRewardBlock() const;
-
-    //!
-    //! \brief Get the hash of the last block with a research reward earned by
-    //! the account.
-    //!
-    //! \return The SHA256 hash of the account's last reward block.
-    //!
-    uint256 LastRewardBlockHash() const;
-
-    //!
-    //! \brief Get the height of the last block with a research reward earned
-    //! by the account.
-    //!
-    //! \return A block height of zero if the account never earned a research
-    //! reward before.
-    //!
-    uint32_t LastRewardHeight() const;
-
-    //!
-    //! \brief Get the timestamp of the last block with a research reward earned
-    //! by the account.
-    //!
-    //! \return A timestamp of zero if the account never earned a research
-    //! reward before.
-    //!
-    int64_t LastRewardTime() const;
-
-    //!
-    //! \brief Get the magnitude of the CPID in the last block with a research
-    //! reward earned by the account.
-    //!
-    //! \return A magnitude of zero if the account never earned a research
-    //! reward before.
-    //!
-    uint16_t LastRewardMagnitude() const;
-
-    //!
-    //! \brief Get the average magnitude of the CPID over every block with a
-    //! research reward earned by the account.
-    //!
-    //! \return A magnitude of zero if the account never earned a research
-    //! reward before.
-    //!
-    double AverageLifetimeMagnitude() const;
-};
-
-//!
-//! \brief A traversable range of all the research accounts stored in the tally.
-//!
-//! This type provides \c const access to the research accounts in the tally
-//! without exposing the internal storage or implementation.
-//!
-class ResearchAccountRange
-{
-    typedef std::unordered_map<Cpid, ResearchAccount> StorageType;
-
-public:
-    typedef StorageType::const_iterator const_iterator;
-    typedef StorageType::size_type size_type;
-
-    //!
-    //! \brief Initialze the wrapper.
-    //!
-    //! \param accounts Points to the accounts stored in the tally.
-    //!
-    ResearchAccountRange(const StorageType& accounts) : m_accounts(accounts)
-    {
-    }
-
-    //!
-    //! \brief Returns an iterator to the beginning.
-    //!
-    const_iterator begin() const
-    {
-        return m_accounts.begin();
-    }
-
-    //!
-    //! \brief Returns an iterator to the end.
-    //!
-    const_iterator end() const
-    {
-        return m_accounts.end();
-    }
-
-    //!
-    //! \brief Get the number of accounts in the tally.
-    //!
-    size_type size() const
-    {
-        return m_accounts.size();
-    }
-
-private:
-    const StorageType& m_accounts; //!< The accounts stored in the tally.
-};
 
 //!
 //! \brief A calculator that computes the accrued research rewards for a
